[DC] Move role assignment calls into AuthService
